ui reducer and actions misplacement fixed
